requirements.txt
django-cms>=2.4.1
